+++ test-types/minimal-openapiV3-document.ts
