+++ test-types/swagger-ui-vendor-extensions.test-d.ts
